requirements.txt
Django==1.10
django-extensions==1.7.1
djangorestframework==3.4.4
psycopg2==2.6.2
Pygments==2.1.3
six==1.10.0
tqdm==4.8.3
